backend/requirements.txt
Flask==3.1.1
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
web3==7.14.1
python-dotenv==1.1.0
scikit-learn==1.6.1
Pillow==11.1.0
imagehash==4.3.2
Werkzeug==3.1.3
firebase-admin>=6.0
sentence-transformers>=3.0
torch>=2.0
torchvision>=0.15
beautifulsoup4>=4.12
requests>=2.31
pinatapy-vourhey>=0.2
